+++ common/CMakeLists.txt
@@ -19,6 +19,8 @@ set(COMMON_SRC
         "${CMAKE_CURRENT_LIST_DIR}/opengl3.h"
         "${CMAKE_CURRENT_LIST_DIR}/rs-config.h"
         "${CMAKE_CURRENT_LIST_DIR}/rs-config.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/fs.h"
+        "${CMAKE_CURRENT_LIST_DIR}/fs.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/os.h"
         "${CMAKE_CURRENT_LIST_DIR}/os.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/fw-update-helper.h"
