+++ examples/record-playback/CMakeLists.txt
@@ -5,11 +5,13 @@ cmake_minimum_required(VERSION 3.1.0)
 
 project(RealsenseExamplesRecord-Playback)
 
+
 if(BUILD_GRAPHICAL_EXAMPLES)
-    add_executable(rs-record-playback rs-record-playback.cpp ../example.hpp ../../third-party/imgui/imgui.cpp ../../third-party/imgui/imgui_draw.cpp ../../third-party/imgui/imgui_impl_glfw.cpp)
+    add_executable(rs-record-playback rs-record-playback.cpp ../example.hpp ../../third-party/imgui/imgui.cpp ../../third-party/imgui/imgui_draw.cpp ../../third-party/imgui/imgui_impl_glfw.cpp ../../common/fs.cpp)
     set_property(TARGET rs-record-playback PROPERTY CXX_STANDARD 11)
     target_link_libraries(rs-record-playback ${DEPENDENCIES})
     include_directories(../ ../../third-party/tclap/include ../../third-party/imgui)
+    include_directories(../../common)
     set_target_properties (rs-record-playback PROPERTIES FOLDER "Examples")
     install(TARGETS rs-record-playback RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR})
 endif()
